# Implement `vec_assign(slice_value = TRUE)` and optimized logical `i` handling