--- a/R/slice.R
+++ b/R/slice.R
@@ -1,21 +1,33 @@
 #' Get or set observations in a vector
 #'
 #' This provides a common interface to extracting and modifying observations
-#' for all vector types, regardless of dimensionality. It is an analog to `[`
-#' that matches [vec_size()] instead of `length()`.
+#' for all vector types, regardless of dimensionality. They are analogs to `[`
+#' and `[<-` that match [vec_size()] instead of `length()`.
 #'
 #' @inheritParams rlang::args_dots_empty
 #' @inheritParams rlang::args_error_context
 #'
 #' @param x A vector
-#' @param i An integer, character or logical vector specifying the
-#'   locations or names of the observations to get/set. Specify
-#'   `TRUE` to index all elements (as in `x[]`), or `NULL`, `FALSE` or
-#'   `integer()` to index none (as in `x[NULL]`).
-#' @param value Replacement values. `value` is cast to the type of
-#'   `x`, but only if they have a common type. See below for examples
-#'   of this rule. `value` must be recyclable to the size of `i` after `i`
-#'   has been converted to a valid integer location vector.
+#'
+#' @param i An integer, character or logical vector specifying the locations or
+#'   names of the observations to get/set. Specify `TRUE` to index all elements
+#'   (as in `x[]`), or `NULL`, `FALSE` or `integer()` to index none (as in
+#'   `x[NULL]`).
+#'
+#' @param value A vector of replacement values
+#'
+#'   `value` is cast to the type of `x`.
+#'
+#'   If `slice_value = FALSE`, `value` must be size 1 or the same size as `i`
+#'   after `i` has been converted to a positive integer location vector with
+#'   [vec_as_location()] (which may not be the same size as `i` originally).
+#'
+#'   If `slice_value = TRUE`, `value` must be size 1 or the same size as `x`.
+#'
+#' @param slice_value A boolean. If `TRUE`, the assignment proceeds as if you
+#'   had provided `vec_slice(x, i) <- vec_slice(value, i)`, but is optimized to
+#'   avoid materializing the slice of `value`.
+#'
 #' @param x_arg,value_arg Argument names for `x` and `value`. These are used
 #'   in error messages to inform the user about the locations of
 #'   incompatible types and sizes (see [stop_incompatible_type()] and
@@ -119,6 +131,31 @@
 #' # vector:
 #' x <- 1:3
 #' try(vec_slice(x, 2) <- 1.5)
+#'
+#' # Slicing `value` -------------------------------------------------
+#'
+#' # Sometimes both `x` and `value` start from objects that are the same length,
+#' # and you need to slice `value` by `i` before assigning it to `x`. This comes
+#' # up when thinking about how `base::ifelse()` and `dplyr::case_when()` work.
+#' condition <- c(TRUE, FALSE, TRUE, FALSE)
+#' yes <- 1:4
+#' no <- 5:8
+#'
+#' # Create an output container and fill it
+#' out <- vec_init(integer(), 4)
+#' out <- vec_assign(out, condition, vec_slice(yes, condition))
+#' out <- vec_assign(out, !condition, vec_slice(no, !condition))
+#' out
+#'
+#' # This is wasteful because you have to materialize the slices of `yes` and
+#' # `no` before they can be assigned, and you also have to validate `condition`
+#' # multiple times. Using `slice_value` internally performs
+#' # `vec_slice(yes, condition)` and `vec_slice(no, !condition)` for you,
+#' # but does so in a way that avoids the materialization.
+#' out <- vec_init(integer(), 4)
+#' out <- vec_assign(out, condition, yes, slice_value = TRUE)
+#' out <- vec_assign(out, !condition, no, slice_value = TRUE)
+#' out
 vec_slice <- function(x, i, ..., error_call = current_env()) {
   check_dots_empty0(...)
   .Call(ffi_slice, x, i, environment())
@@ -188,20 +225,28 @@ vec_slice_dispatch_integer64 <- function(x, i) {
 `vec_slice<-` <- function(x, i, value) {
   x_arg <- "" # Substitution is `*tmp*`
   delayedAssign("value_arg", as_label(substitute(value)))
-
-  .Call(ffi_assign, x, i, value, environment())
+  slice_value <- FALSE
+  .Call(ffi_assign, x, i, value, slice_value, environment())
 }
 #' @rdname vec_slice
 #' @export
-vec_assign <- function(x, i, value, ..., x_arg = "", value_arg = "") {
+vec_assign <- function(
+  x,
+  i,
+  value,
+  ...,
+  slice_value = FALSE,
+  x_arg = "",
+  value_arg = ""
+) {
   check_dots_empty0(...)
-  .Call(ffi_assign, x, i, value, environment())
+  .Call(ffi_assign, x, i, value, slice_value, environment())
 }
 
 # Invariants for `[<-` methods:
 #
 # - `i` will contain positive integers
-# - `i` will not contain `NA` (a base `[<-` bug gets in the way if we allow this)
+# - `i` will not contain `NA` (a base `[<-` issue gets in the way if we allow this)
 # - `value` will be size 1 or size `length(i)`
 #
 # Given these invariants, the base `[<-` works the way we want it to in fallback
@@ -209,12 +254,31 @@ vec_assign <- function(x, i, value, ..., x_arg = "", value_arg = "") {
 # `[<-` with these inputs. In other words, we don't expect subclasses to have
 # vctrs subassign behavior, but we do expect them to match a subset of base R
 # subassign behavior.
-vec_assign_fallback <- function(x, i, value) {
-  # Work around bug in base `[<-`
+vec_assign_fallback <- function(x, i, value, slice_value, index_style) {
+  if (index_style == "condition") {
+    # Convert logical condition `i` to integer location `i`. Must use
+    # `vec_as_location()` rather than `which()` because we want `NA` values to
+    # propagate, which is consistent with how `value`'s size is checked
+    # internally (i.e. when `slice_value = FALSE`, `value` was checked to have
+    # size equal to the number of `TRUE` or `NA` values in `i`). Propagated
+    # `NA`s are later dropped before calling `[<-` to work around a base `[<-`
+    # issue, but we need them to slice both `i` and `value` consistently.
+    i <- vec_as_location(i, n = vec_size(x), missing = "propagate")
+  }
+
+  if (slice_value && vec_size(value) != 1L) {
+    # `value` has same size as `x` rather than same size as `i`.
+    # We need to pre-slice it down to same size as `i` to match what `[<-` expects.
+    # Effectively we are preemptively doing the RHS of this:
+    # vec_slice(x, i) <- vec_slice(value, i)
+    value <- vec_slice(value, i)
+  }
+
+  # Work around issue in base `[<-` that errors on `NA_integer_` in subassign
+  # indices
   existing <- !is.na(i)
   i <- vec_slice(i, existing)
   if (vec_size(value) != 1L) {
-    # Only slice `value` if we aren't recycling it
     value <- vec_slice(value, existing)
   }
 
@@ -225,12 +289,41 @@ vec_assign_fallback <- function(x, i, value) {
 }
 
 # `start` is 0-based
-vec_assign_seq <- function(x, value, start, size, increasing = TRUE) {
-  .Call(ffi_assign_seq, x, value, start, size, increasing)
+vec_assign_seq <- function(
+  x,
+  value,
+  start,
+  size,
+  increasing = TRUE,
+  slice_value = FALSE
+) {
+  .Call(ffi_assign_seq, x, value, start, size, increasing, slice_value)
 }
 
-vec_assign_params <- function(x, i, value, assign_names = FALSE) {
-  .Call(ffi_assign_params, x, i, value, assign_names)
+#' @param assign_names A boolean. If `TRUE`, will assign names from `value`
+#'   onto `x` as well.
+#'
+#' @noRd
+vec_assign_params <- function(
+  x,
+  i,
+  value,
+  ...,
+  assign_names = FALSE,
+  slice_value = FALSE,
+  x_arg = "",
+  value_arg = ""
+) {
+  check_dots_empty0(...)
+  .Call(
+    ffi_assign_params,
+    x,
+    i,
+    value,
+    assign_names,
+    slice_value,
+    environment()
+  )
 }
 
 vec_remove <- function(x, i) {