--- a/man/vec_as_location.Rd
+++ b/man/vec_as_location.Rd
@@ -50,10 +50,7 @@ num_as_location2(
 )
 }
 \arguments{
-\item{i}{An integer, character or logical vector specifying the
-locations or names of the observations to get/set. Specify
-\code{TRUE} to index all elements (as in \code{x[]}), or \code{NULL}, \code{FALSE} or
-\code{integer()} to index none (as in \code{x[NULL]}).}
+\item{i}{An index vector to convert.}
 
 \item{n}{A single integer representing the total size of the
 object that \code{i} is meant to index into.}