--- a/src/c.c
+++ b/src/c.c
@@ -85,6 +85,8 @@ r_obj* vec_c_opts(r_obj* xs,
   const struct vec_proxy_assign_opts c_proxy_assign_opts = {
     .ownership = VCTRS_OWNERSHIP_deep,
     .recursively_proxied = true,
+    .slice_value = ASSIGNMENT_SLICE_VALUE_no,
+    .index_style = VCTRS_INDEX_STYLE_location,
     .assign_names = assign_names,
     .ignore_outer_names = true,
     .call = error_call
@@ -134,7 +136,14 @@ r_obj* vec_c_opts(r_obj* xs,
         // If there is no name to assign, skip the assignment since
         // `out_names` already contains empty strings
         if (x_nms != chrs_empty) {
-          out_names = chr_assign(out_names, loc, x_nms, VCTRS_OWNERSHIP_deep);
+          out_names = chr_assign(
+            out_names,
+            loc,
+            x_nms,
+            VCTRS_OWNERSHIP_deep,
+            ASSIGNMENT_SLICE_VALUE_no,
+            VCTRS_INDEX_STYLE_location
+          );
           KEEP_AT(out_names, out_names_pi);
         }
       }