… NEWS.md
… src/slice-assign-array.c
… src/slice-assign.c
… src/utils.c
… src/utils.h
… src/vctrs-core.h
… tests/testthat/_snaps/slice-assign.md
… tests/testthat/test-slice-assign.R