Two improvements in this PR

- `assignment_index_style` which allows for either integer _location_ indices or logical _condition_ indices. Each have their own optimized path inside of `vec_assign()`. This is auto detected for the user in `vec_assign()`. If it doesn't look like a simple logical condition index, we run it through `vec_as_location()` and use the location path. The logical condition index path is the new one.

- `assignment_slice_value` which says whether or not to optionally slice `value` by `i` before assignment as well.
    - If no, the default, assignment is done as expected, `x[i] <- value`.
    - If yes, assignment is done as an optimized form of `x[i] <- value[i]`, where `value[i]` is not actually materialized internally.
    - This is exposed as `vec_assign(slice_value = FALSE/TRUE)`

Note that `slice_value = TRUE` works regardless of the index style. These are orthogonal ideas. Figuring this out was a key part of this PR.

Taken together, this allows for optimal assignment in `dplyr::if_else()` and `dplyr::case_when()` style assignments where you have:
- LHSs that are logical condition indices, which we can now assign with directly rather than doing `which(i)` first
- RHSs that are length 1 or the length of the output, which can use `slice_value = TRUE` during assignment

This should allow for creation of the output with minimal extra allocations and overhead.

In a follow up PR I will create `list_multi_assign()`, which will eventually benefit from these optimizations and expose `slice_values = FALSE/TRUE`, and `vec_case_when()` and `vec_if_else()` will use that function directly. But getting these core assign details out of the way is the hard part.

---

Here's an example of the intended use case. `vec_assign()` with a logical index and a `value` that can benefit from `slice_value = TRUE`. This would be used `N` times by `dplyr::case_when()` where `N` is the number of provided cases.

``` r
library(vctrs)
set.seed(123)

x <- sample(1e7)
value <- sample(1e7)

# When `x > 100000`, put `value`, otherwise keep `x`
loc <- x > 100000

bench::mark(
  old = {
    # Improvements in `vec_assign()` logical `loc` handling
    vec_assign(x, loc, vec_slice(value, loc))
  },
  new = {
    # Additional optimization with `slice_value`
    vec_assign(x, loc, value, slice_value = TRUE)
  }
)
#> # A tibble: 2 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old          26.8ms   27.5ms      36.0   113.7MB     36.0
#> 2 new           7.1ms   7.49ms     114.     38.1MB     26.0

# This is CRAN vctrs with no optimizations
bench::mark(
  cran = {
    vec_assign(x, loc, vec_slice(value, loc))
  },
  iterations = 50
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 cran         35.9ms   36.8ms      26.8     151MB     27.3
```

``` r
profmem::profmem(vec_assign(x, loc, value, slice_value = TRUE))
#> Rprofmem memory profiling of:
#> vec_assign(x, loc, value, slice_value = TRUE)
#> 
#> Memory allocations:
#> Number of 'new page' entries not displayed: 3
#>        what    bytes        calls
#> 3     alloc 40000048 vec_assign()
#> total       40000048
```

Here's an even more compelling example with a data frame, which has multiple columns so the overhead of slicing it is higher.

```r
library(vctrs)
set.seed(123)

dictionary <- data.frame(a = 1:10, b = 1:10, c = 1:10 + 0, d = letters[1:10])

x <- dplyr::slice_sample(dictionary, n = 1e7, replace = TRUE)
value <- dplyr::slice_sample(dictionary, n = 1e7, replace = TRUE)

# When `sample(1e7) > 100000`, put `value`, otherwise keep `x`
loc <- sample(1e7) > 100000

# This PR
bench::mark(
  old = {
    # Improvements in `vec_assign()` logical `loc` handling
    vec_assign(x, loc, vec_slice(value, loc))
  },
  new = {
    # Additional optimization with `slice_value`
    vec_assign(x, loc, value, slice_value = TRUE)
  },
  iterations = 20
)
#> # A tibble: 2 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old         129.5ms  130.7ms      7.65     493MB     29.1
#> 2 new          64.4ms   65.1ms     15.3      229MB     10.2


# This is CRAN vctrs with no optimizations
bench::mark(
  cran = {
    vec_assign(x, loc, vec_slice(value, loc))
  },
  iterations = 50
)
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 cran          147ms    148ms      6.75     531MB     169.
```